// apps/readest-app/src/types/book.ts
export type BookFormat = 'EPUB' | 'PDF' | 'MOBI' | 'CBZ' | 'FB2' | 'FBZ';
export type BookNoteType = 'bookmark' | 'annotation' | 'excerpt';
export type HighlightStyle = 'highlight' | 'underline' | 'squiggly';
export type HighlightColor = 'red' | 'yellow' | 'green' | 'blue' | 'violet';

export interface Book {
    // if Book is a remote book we just lazy load the book content via url
    url?: string;
    // if Book is a transient local book we can load the book content via filePath
    filePath?: string;
    hash: string;
    format: BookFormat;
    title: string;
    author: string;
    group?: string; // deprecated in favor of groupId and groupName
    groupId?: string;
    groupName?: string;
    tags?: string[];
    coverImageUrl?: string | null;

    createdAt: number;
    updatedAt: number;
    deletedAt?: number | null;

    uploadedAt?: number | null;
    downloadedAt?: number | null;

    lastUpdated?: number; // deprecated in favor of updatedAt
    progress?: [number, number]; // Add progress field: [current, total], 1-based page number
    primaryLanguage?: string;
}